+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## v0.0.3 - 2020-04-26
+### Removed
+- Remove `mongoose` folder
+- Remove `index.js` file (see **CHANGED**)
+
+### Changed
+- Rename `index2.js` to `index.js`
+
 ## v0.0.2 - 2020-04-26
 ### Added
 - CHANGELOG.md file